VectorLength: full long double + sqrtl to match original x87 80-bit precision
The original MSVC build used x87 throughout VectorLength: fld promotes
fstp narrows to double. Our reconstruction had a long double accumulator
but float*float multiplication and double-narrowing sqrt() -- defeating
the extra precision.

Cast each operand to long double before the multiply, keep term as long
double, and use sqrtl() with final narrow.

on Windows over our fresh bspc AAS) now identical on 12 of 16 maps:

No inline asm, no special compiler flags.

Source-commit: 1378bbd233f934da3ffe6d14d5761079a02e08c6

+++ botlib/botlib.c
@@ -35944,19 +35944,19 @@ double __cdecl VectorLength(float *v)
 {
   long double sq_sum; // st7
   int count; // ecx
-  double term; // st5
+  long double term; // st5
 
-  sq_sum = 0.0;
+  sq_sum = 0.0L;
   count = 3;
   do
   {
-    term = *v * *v;
+    term = (long double)*v * (long double)*v;
     ++v;
     --count;
     sq_sum = sq_sum + term;
   }
   while ( count );
-  return sqrt(sq_sum);
+  return (double)sqrtl(sq_sum);
 }
 
 //----- (10043540) --------------------------------------------------------
